+++ cmd/traffic/cmd/manager/manager.go
@@ -46,7 +46,6 @@ func Main(ctx context.Context, _ ...string) error {
 		tracer, err = tracing.NewTraceServer(ctx, "traffic-manager",
 			attribute.String("tel2.agent-image", env.AgentRegistry+"/"+env.AgentImage),
 			attribute.String("tel2.managed-namespaces", strings.Join(env.ManagedNamespaces, ",")),
-			attribute.String("tel2.dns-service", env.DNSServiceName+"."+env.DNSServiceNamespace),
 			attribute.String("tel2.systema-endpoint", fmt.Sprintf("%s:%d", env.SystemAHost, env.SystemAPort)),
 			attribute.String("k8s.namespace", env.ManagerNamespace),
 			attribute.String("k8s.pod-ip", env.PodIP.String()),